DEV - removed for_each as its only supported in MSVC.
Updated unsigned int warnings.

+++ src/two_sum/two_sum.cpp
@@ -10,13 +10,13 @@ vector<int> TwoSum::find(const vector<int>& nums, int target) {
     unordered_map<int, int> saved_diff;
     saved_diff.reserve(nums.size());
     
-    for (int i = 0; i < nums.size(); ++i)
+    for (unsigned int i = 0; i < nums.size(); ++i)
     {
         int diff = target - nums[i];
         auto iter = saved_diff.find(diff);
         
         if (iter != saved_diff.end())
-            return vector<int>{ iter->second, i };
+            return vector<int>{ iter->second, (int)i };
             
         saved_diff[nums[i]] = i;   
     }
